Workaround for macros defined by Cortex M4 ('core_cm4.h') conflicting with identifiers internally used by Gcc5 (issue #1).
CMake will check whether the workaround is required. If so, the
conflicting macros are undef'ed. Files should therefore not include any
Stm32 files directly, but the dedicated Platform.h, which takes care
of this issue.

Target builds (= crosscompiling) will not check for the workaround and
should fail in such a case since that indicates a severe integration
failure.